changed naming to _internal_t scheme in docgen_worker.hpp to prevent name conflicts; moved Token<Symbol> specializations from docgen_worker.hpp to token.hpp
