+++ src/core/parser_internal/docgen_worker.hpp
@@ -10,32 +10,16 @@ namespace docgen {
 namespace core {
 namespace parser_internal {
 
-using token_class_t = Token<Symbol>;
+using symbol_internal_t = Symbol;
 
-using writer_class_t = JSONWriter<token_class_t>;
+using token_internal_t = Token<symbol_internal_t>;
 
-using worker_class_t = ParseWorker<token_class_t, writer_class_t>;
+using writer_internal_t = JSONWriter<token_internal_t>;
 
-using routine_details_t = WorkerRoutine<worker_class_t>;
+using worker_internal_t = ParseWorker<token_internal_t, writer_internal_t>;
+
+using routine_details_internal_t = WorkerRoutine<worker_internal_t>;
 
 } // namespace parser_internal
-
-inline bool operator==(const Token<Symbol>& t1, const Token<Symbol>& t2) {
-	return t1.name == t2.name;
-}
-
 } // namespace core
 } // namespace docgen
-
-namespace std {
-
-template <>
-struct hash<docgen::core::Token<docgen::core::Symbol>>
-{
-	size_t operator()(const docgen::core::Token<docgen::core::Symbol>& t) const
-	{
-		return hash<size_t>()(static_cast<size_t>(t.name));
-	}
-};
-
-} // namespace std
